added query script. Not a container, per se, but a set of functions that constructs the necessary strings, removing that duty from the database class.
