+++ clinical_study_organizer/database.py
@@ -1,5 +1,5 @@
 import sqlite3 as db
-
+from clinical_study_organizer.containers.query import *
 
 # Written with help of https://docs.python.org/2/library/sqlite3.html
 from clinical_study_organizer.containers.query_result import Query_Result
@@ -12,73 +12,17 @@ class Database:
         self.cursor = None
 
     def initialize(self, attribute_dictionary):
-        attribute_string = self._construct_attribute_table(attribute_dictionary)
-
-        identity_table = "CREATE TABLE identity (" \
-                         "alias      VARCHAR REFERENCES alias (alias),\n" \
-                         "id         INT UNIQUE PRIMARY KEY,\n" \
-                         "last_name         VARCHAR,\n" \
-                         "first_name         VARCHAR" \
-                         ");"
-
-        attribute_table = "CREATE TABLE attributes (\n" \
-                          "alias    VARCHAR PRIMARY KEY UNIQUE,\n" + \
-                          attribute_string + \
-                          ");"
-
+        attribute_table, identity_table = get_initial_tables(attribute_dictionary)
         self._execute_SQL([identity_table, attribute_table])
 
     def delete_tables(self):
         self._execute_SQL(["DROP TABLE attributes;", "DROP TABLE identity;"])
 
-
-    def _execute_SQL(self, commands):
-        self._open()
-        for command in commands:
-            self.cursor.execute(command)
-        self._close()
-
-    def _query_database(self, query):
-        self._open()
-        self.cursor.execute(query)
-        info = self.cursor.fetchall()
-        self._close()
-
-        return info
-
-
-    def _open(self):
-        self.connection = db.connect(self.database_name)
-        self.cursor = self.connection.cursor()
-
-    def _close(self):
-        self.connection.commit()
-        self.connection.close()
-
     def add_patients_attributes(self, patients):
-        queries = []
-
-        for patient in patients:
-            alias = "\'" + patient.alias + "\', "
-            data = patient.data
-            data_string = self._construct_patient_attributes(data)
-
-            query = "INSERT INTO attributes VALUES (" + \
-            alias + \
-            data_string + \
-            ");"
-
-            queries.append(query)
-
-        self._execute_SQL(queries)
+        self._execute_SQL(construct_patients_attributes(patients))
 
     def add_patients_identities(self, patients):
-        queries = []
-
-        for patient in patients:
-            query = "INSERT INTO identity VALUES (\'" + patient.alias + "\', \'" + patient.id + "\', \'" + patient.last_name + "\', \'" + patient.first_name + "\')"
-            queries.append(query)
-
+        queries = construct_patients_identities(patients)
         self._execute_SQL(queries)
 
     def get_data(self, alias):
@@ -94,51 +38,32 @@ class Database:
         return self._query_database(query)
 
     def get_all_attribute_values(self, attributes):
-        query = "SELECT alias, " + self._construct_attributes_select(attributes) + " FROM attributes;"
+        query = construct_all_attribute_values(attributes)
 
         return Query_Result(self._query_database(query), attributes)
-        
-    @staticmethod
-    def _construct_patient_attributes(data):
-        patient_data = []
-        
-        for value in data:
-            patient_data.append("\'" + str(value) + "\', ")
 
-        remove_last_comma(patient_data)
-        
-        return "".join(patient_data)
+    def _execute_SQL(self, commands):
+        self._open()
+        for command in commands:
+            self.cursor.execute(command)
+        self._close()
 
-    @staticmethod
-    def _construct_attribute_table(attribute_dictionary):
-        attribute_string_list = []
+    def _query_database(self, query, attributes=None):
+        self._open()
+        self.cursor.execute(query)
+        info = self.cursor.fetchall()
+        self._close()
 
-        for attribute_name in (attribute for attribute in attribute_dictionary if
-                               attribute != "id" and attribute != "last_name" and attribute != "first_name"):
-            attribute_type = attribute_dictionary[attribute_name]
-            attribute_string = attribute_name + " " + attribute_type + ",\n"
+        return info
 
-            attribute_string_list.append(attribute_string)
+    def _open(self):
+        self.connection = db.connect(self.database_name)
+        self.cursor = self.connection.cursor()
 
-        remove_last_comma(attribute_string_list)
+    def _close(self):
+        self.connection.commit()
+        self.connection.close()
 
-        return "".join(attribute_string_list)
-
-    @staticmethod
-    def _construct_attributes_select(attributes):
-        query = []
-
-        for attribute in attributes:
-            line = attribute + ", "
-            query.append(line)
-
-        remove_last_comma(query)
-
-        return "".join(query)
-
-
-def remove_last_comma(string_list):
-    string_list[len(string_list) - 1] = string_list[len(string_list) - 1].replace(",", "")
