+++ test/three_laws_of_TDD.md
@@ -1,0 +1,11 @@
+# Three Laws of TDD
+Taken from Robert C. Martin's *Clean Code*
+
+## 1st Law
+You may not write production code until you have written a failing unit test.
+
+## 2nd Law
+You may not write more of a unit test than is sufficient to fail, and not compiling is failing.
+
+## 3rd Law
+You may not write more production code than is sufficient to pass the currently failing test.
